medical_llm/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-llm-engine"
version = "1.0.0"
description = "Multi-model medical LLM engine with RAG, safety validation, and structured reporting"
requires-python = ">=3.9"
license = {text = "Research Use Only"}
readme = "README.md"

dependencies = [
    # Core ML Frameworks
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "safetensors>=0.4.0",
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",
    "tokenizers>=0.15.0",

    # Embeddings & Vector Store
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.3.0",

    # NLP & NER
    "scispacy>=0.5.3",
    "spacy>=3.7.0",

    # RAG & Search
    "duckduckgo-search>=5.0.0",
    "requests>=2.31.0",
    "xmltodict>=0.13.0",
    "trafilatura>=1.9.0",
    "pypdf>=4.2.0",
    "qdrant-client>=1.11.0",
    "langgraph>=0.2.34",

    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.65.0",

    # API Server
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "redis[hiredis]>=5.0.0",

    # Logging & Monitoring
    "rich>=13.7.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.4.0"]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
medical-llm = "main:main"

[tool.setuptools]
py-modules = ["app", "main"]
include-package-data = true

[tool.setuptools.packages.find]
include = ["core*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py39"
line-length = 100
